requirements.txt
Django==1.9
Pillow==4.0.0
django-sortedm2m==1.2.2
